Add custom masonry system
CSS Grid masonry is experimental and unsuitable for usage on real
websites at this time. Let's get rid of it.

Tested out most of the approaches outlined here, as well as the
libraries.

https://css-tricks.com/piecing-together-approaches-for-a-css-masonry-layout/

None of these really work for images with arbitrary sizes while
allowing adding of content dynamically without causing jumping in
the layout.

Not that I could figure out within a reasonable time anyway...

Because of this I created a custom masonry system based on N columns
of flexboxes, where images are continually fed into the column with
the least content.

Since images only have their minimum size of 150px when they get added
to the dom and only get their real size after they are downloaded it
also does reflowing of content whenever an image loads.